improved App.js code styling
